requirements.txt
asgiref==3.2.7
Django==3.0.7
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
factory-boy==2.12.0
Faker==4.1.0
gunicorn==20.0.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
